+++ hypnotoad/scripts/hypnotoad_plot_grid_cells.py
@@ -136,27 +136,28 @@ def main():
         # Note: have to use `s` to set marker size, because xarray has used a `size`
         # kwarg for something else.
         ds_region.isel(
-            x=slice(xin, xout_centre), theta=slice(ylow, yup_centre)
-        ).plot.scatter("R", "Z", marker=".", color="k", s=1)
+            x=slice(xin, xout_centre),
+            y=slice(ylow, yup_centre),
+        ).plot.scatter(x="R", y="Z", marker=".", color="k", s=1)
 
         # plot radial grid lines
         plt.plot(
-            ds_region["Rxy_corners"].isel(
-                x=slice(xin, xout_corner_rad), theta=slice(ylow, yup_corner)
+            ds_region["Rxy_lower_left_corners"].isel(
+                x=slice(xin, xout_corner_rad), y=slice(ylow, yup_corner)
             ),
-            ds_region["Zxy_corners"].isel(
-                x=slice(xin, xout_corner_rad), theta=slice(ylow, yup_corner)
+            ds_region["Zxy_lower_left_corners"].isel(
+                x=slice(xin, xout_corner_rad), y=slice(ylow, yup_corner)
             ),
             color="k",
         )
 
         # plot poloidal grid lines
         plt.plot(
-            ds_region["Rxy_corners"]
-            .isel(x=slice(xin_pol, xout_corner_pol), theta=slice(ylow, yup_corner))
+            ds_region["Rxy_lower_left_corners"]
+            .isel(x=slice(xin_pol, xout_corner_pol), y=slice(ylow, yup_corner))
             .T,
-            ds_region["Zxy_corners"]
-            .isel(x=slice(xin_pol, xout_corner_pol), theta=slice(ylow, yup_corner))
+            ds_region["Zxy_lower_left_corners"]
+            .isel(x=slice(xin_pol, xout_corner_pol), y=slice(ylow, yup_corner))
             .T,
             color="k",
         )
@@ -175,8 +176,12 @@ def main():
             # neighbouring in the global grid, because if they are the boundary between
             # regions is not (or 'not really') a branch cut.
             plt.plot(
-                ds_region["Rxy_corners"].isel(x=slice(xin, xout_corner_rad), theta=-1),
-                ds_region["Zxy_corners"].isel(x=slice(xin, xout_corner_rad), theta=-1),
+                ds_region["Rxy_lower_left_corners"].isel(
+                    x=slice(xin, xout_corner_rad), y=-1
+                ),
+                ds_region["Zxy_lower_left_corners"].isel(
+                    x=slice(xin, xout_corner_rad), y=-1
+                ),
                 color="r",
                 linewidth=3,
                 zorder=1000,
@@ -187,8 +192,12 @@ def main():
             # By arbitrary choice, plot from the region(s) inside the separatrix, so
             # highlight the outer edge.
             plt.plot(
-                ds_region["Rxy_corners"].isel(x=-1, theta=slice(ylow, yup_corner)),
-                ds_region["Zxy_corners"].isel(x=-1, theta=slice(ylow, yup_corner)),
+                ds_region["Rxy_lower_left_corners"].isel(
+                    x=-1, y=slice(ylow, yup_corner)
+                ),
+                ds_region["Zxy_lower_left_corners"].isel(
+                    x=-1, y=slice(ylow, yup_corner)
+                ),
                 color="b",
                 linewidth=3,
                 zorder=999,
@@ -197,11 +206,11 @@ def main():
         if targets:
             if ds_region.regions[r].connection_lower_y is None:
                 plt.plot(
-                    ds_region["Rxy_corners"].isel(
-                        x=slice(xin, xout_corner_rad), theta=ylow
+                    ds_region["Rxy_lower_left_corners"].isel(
+                        x=slice(xin, xout_corner_rad), y=ylow
                     ),
-                    ds_region["Zxy_corners"].isel(
-                        x=slice(xin, xout_corner_rad), theta=ylow
+                    ds_region["Zxy_lower_left_corners"].isel(
+                        x=slice(xin, xout_corner_rad), y=ylow
                     ),
                     color="k",
                     linewidth=3,
@@ -210,11 +219,11 @@ def main():
             if ds_region.regions[r].connection_upper_y is None:
                 yval = -1 if yup_corner is None else yup_corner
                 plt.plot(
-                    ds_region["Rxy_corners"].isel(
-                        x=slice(xin, xout_corner_rad), theta=yval
+                    ds_region["Rxy_lower_left_corners"].isel(
+                        x=slice(xin, xout_corner_rad), y=yval
                     ),
-                    ds_region["Zxy_corners"].isel(
-                        x=slice(xin, xout_corner_rad), theta=yval
+                    ds_region["Zxy_lower_left_corners"].isel(
+                        x=slice(xin, xout_corner_rad), y=yval
                     ),
                     color="k",
                     linewidth=3,