+++ doc/whats-new.md
@@ -2,6 +2,13 @@ Release history
 ===============
 
 ### Bug fixes
+- When using the GUI, if there is an error in `TokamakEquilibrium` object
+  creation, still plot the equilibrium data (#186).
+  By [John Omotani](https://github.com/johnomotani)
+- Set the dimension for R_closed_wall and Z_closed_wall to 'closed_wall'. Fixes
+  loading of grid files by xBOUT (#191).
+  By [John Omotani](https://github.com/johnomotani)
+- Calculate the parallel distance along field lines, and save to the output (#193).
 
 
 ### New features
@@ -9,12 +16,25 @@ Release history
 - Radial grid line construction can recover from failure and generate
   a rough grid to help visual inspection when option
   `follow_perpendicular_recover` is set to True (#175)
+  By [Ben Dudson](https://github.com/bendudson)
 - A `View` menu enables the grid plot to be customised, with cell edges, corners,
   grid lines and other components (#176).
+  By [Ben Dudson](https://github.com/bendudson)
 - `penalty_mask` is calculated and written to the grid file, based on intersection
   of the grid with the wall. This enables immersed boundary conditions.
+  By [Ben Dudson](https://github.com/bendudson)
 - Wall coordinates are written to output grid as `closed_wall_R` and `closed_wall_Z`
   (#176)
+  By [Ben Dudson](https://github.com/bendudson)
+- Extend divertor legs with `leg_extend` options. These specify how far each
+  leg should extend beyond the wall intersection (#195).
+  By [Ben Dudson](https://github.com/bendudson)
+- Linear poloidal spacing option for nonorthogonal grids (#190).
+  By [John Omotani](https://github.com/johnomotani)
+- Grids that are nonorthogonal only at the X-point (#180).
+  By [John Omotani](https://github.com/johnomotani)
+- Script to convert UEDGE grid file to BOUT++ grid file (#136).
+  By [Ben Dudson](https://github.com/bendudson)
 
 0.5.2 (13th March 2023)
 -------------------------